+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## unreleased
 
+### Bugfixes
+- Fix handling mapping that omit key part
+
 ---
 ## 0.8.0 (2024-08-09)